delete type float from 49 line

+++ Lesson2/src/main/java/HomeWork.java
@@ -46,6 +46,6 @@ public class HomeWork {
     }
 
     public static double calculateHypotenuse(int a, int b) {
-        return (float) (Math.sqrt(Math.pow(a, 2) + Math.pow(b, 2)));
+        return Math.sqrt(Math.pow(a, 2) + Math.pow(b, 2));
     }
 }
